Desktop/Scaler School Of technology/procurement_env/requirements.txt
openai>=1.0.0
pyyaml>=6.0
fastapi>=0.110.0
uvicorn>=0.29.0